backend/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.2
pillow>=9.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Utilities
pyyaml>=6.0
matplotlib>=3.7.0
tqdm>=4.65.0

# For advanced features
opencv-python>=4.8.0
scipy>=1.11.0

# Notebooks
ipykernel==6.31.0
jupyterlab==4.4.10